requirements.txt
annotated-types==0.7.0
anyio==4.4.0
blinker==1.8.2
certifi==2024.6.2
click==8.1.7
distro==1.9.0
dnspython==2.6.1
Flask==3.0.3
Flask-Cors==4.0.1
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
install==1.3.5
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
numpy==2.0.0
openai==1.35.2
pydantic==2.7.4
pydantic_core==2.18.4
pymongo==4.7.3
python-dotenv==1.0.1
scipy==1.13.1
sniffio==1.3.1
tenacity==8.4.1
termcolor==2.4.0
tqdm==4.66.4
typing_extensions==4.12.2
Werkzeug==3.0.3
